Fix remaining 84 compilation errors
- Added Add method to DataVector for collection operations
- Fixed unmarshal_float calls to remove incorrect parameters
- Fixed marshal/unmarshal calls to use object methods instead of passing as parameters
- Fixed type casting issues (uint/int conversions)
- Added missing Main class members (txtLog, listItemDesc, images method)
- Fixed ProgressBar.Position to use Value property
- Removed DevExpress .Properties references from Designer files
- Fixed TabControl SelectedTabPage to SelectedIndex
- Fixed CompressionLevel enum casting
- Added missing using GNET directive to WeaponOctetsEditor
- Fixed icon to bitmap conversion
- Changed SkillByClass to use dynamic type for flexibility

+++ C#/pwAdmin - Cliente 2/Protocols/Packets/GRolePocket.cs
@@ -20,7 +20,7 @@ namespace Protocols.Packets
             os.marshal(this.capacity);
             os.marshal(this.timestamp);
             os.marshal(this.money);
-            os.compact_uint32(items.Count);
+            os.compact_uint32((uint)items.Count);
             for (int i = 0; i < items.Count; i++)
             {
                 items[i].marshal(os);
@@ -35,7 +35,7 @@ namespace Protocols.Packets
             this.capacity = os.unmarshal_int();
             this.timestamp = os.unmarshal_int();
             this.money = os.unmarshal_int();
-             int num = os.uncompact_uint32();
+             int num = (int)os.uncompact_uint32();
              for (int i = 0; i < num; i++)
              {
                  GRoleInventory item = new GRoleInventory();
